Latest version of the addon.
This version can import Rigid models and assign textures. Only thing you have to do, is put the textures folder in the addon's folder so it can find it. Or in case you know how to code python, change the import_mef's path to where you have that textures folder.